Annotate the TaskRun when its logs have been copied to GCS
Look for this annotation when determining whether to respond with build
SUCCESS or FAILURE.

gcloud streams logs by polling the logs object in GCS, and
simultaneously polling the GCB API to see whether the build is complete
(SUCCESS, FAILURE, etc.). If the API responds that the build is
complete, gcloud will attempt one more poll of the logs object, stream
it to stdout, then exit.

This causes incomplete logs objects to be shown to gcloud users when the
compat service responds with SUCCESS/FAILURE based only on the status of
the underlying TaskRun -- the TaskRun might be complete, but the logs
aren't necessarily done being copied to GCS, and might not be for a few
more seconds.

In "real" GCB, the worker doesn't report the Build as complete until
it's also finished copying logs, even if all the user's steps have
completed.

To address this issue, the compat service will annotate the underlying
TaskRun with an annotation saying the logs have finished being copied.
When the Build is GETted, during TaskRun->Build conversion, ToBuild will
also check for the presence of the logs-copied annotation when
determining whether it should tell clients that the build is complete.
Only after the client polls while the build is done, and logs are
copied, will the compat service say the build is "done".

This has some downsides:
- meddling users can remove the "logs-copied" annotation, in which case
previously done builds will start to look WORKING again. They can also
modify the status to whatever they want, so it's not that big a deal.
- timing can be misleading. In "real" GCB, a build's finishTime includes
the time it takes to copy logs. The compat service will still respond
with the TaskRun's last step's finishTime, even if it took a few more
seconds to finish writing logs.