Rewrite service as CRD controller
This replaces the K8s Service with a CRD type and Knative-style
controller and webhook admission controller which watches TaskRuns.

This is not currently functional because the Build CRD type (meant to
mimic the GCB Build type) does not have codegen yet to be able to
create/watch/etc those resources on the cluster.

That'll happen next.